requirements.txt
asgiref==3.7.2
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.39
botocore==1.35.39
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
dj-database-url==2.3.0
Django==4.2.16
django-allauth==65.0.2
django-bootstrap-v5==1.0.11
django-bootstrap5==24.3
django-database-url==1.0.3
django-environ==0.11.2
django-pwa==2.0.1
django-storages==1.14.4
django-widget-tweaks==1.5.0
docker==7.1.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
jwt==1.3.1
MarkupSafe==3.0.1
moto==3.1.19
Pillow==10.0.0
packaging==24.2
py==1.11.0
pycparser==2.22
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
s3transfer==0.10.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==2.1.2
whitenoise==6.8.2
xmltodict==0.14.1
